requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
fpdf==1.7.2
idna==3.10
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.3
urllib3==2.5.0
